Optimized and added Macro for repeated task

+++ CuSorting/cuRDStruct.h
@@ -8,7 +8,7 @@ struct PaperIdWrapper {
 };
 
 struct RollNumberWrapper {
-	int rollnumber;
+	long long rollnumber;
 	ResultsDataStructure * classPtr;
 };
